refactor: move display_more_search_results into seperate file
